# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Access information about current playing track on spotify through REST endpoints.",
        "title": "Spoty API",
        "contact": {
            "name": "Jules Michael"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "paths": {
        "/api": {
            "get": {
                "description": "checks if server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "core"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Success"
                        }
                    }
                }
            }
        },
        "/api/authenticate": {
            "get": {
                "description": "redirects user to spotify for authentication",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spoty"
                ],
                "summary": "Authentication",
                "responses": {
                    "302": {
                        "description": "redirection to spotify",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "already authenticated",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    }
                }
            }
        },
        "/api/callback": {
            "get": {
                "description": "spotify redirects to the this endpoint on success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spoty"
                ],
                "summary": "Callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code from spotify",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "state from spotify",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "authenticated successfully",
                        "schema": {
                            "$ref": "#/definitions/server.Success"
                        }
                    },
                    "403": {
                        "description": "could not retrieve token",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    },
                    "404": {
                        "description": "could not retrieve current user",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    }
                }
            }
        },
        "/api/current": {
            "get": {
                "description": "returns information about the current playing track",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spoty"
                ],
                "summary": "Current Playing Track",
                "responses": {
                    "200": {
                        "description": "returns full track information",
                        "schema": {
                            "$ref": "#/definitions/spotify.FullTrack"
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    },
                    "404": {
                        "description": "no current playing track found",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    }
                }
            }
        },
        "/api/current/images": {
            "get": {
                "description": "returns the album images of the current playing track",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spoty"
                ],
                "summary": "Album Images of Current Playing Track",
                "responses": {
                    "200": {
                        "description": "returns album images",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/spoty.Image"
                            }
                        }
                    },
                    "401": {
                        "description": "not authenticated",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    },
                    "404": {
                        "description": "no current playing track found",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    },
                    "500": {
                        "description": "album images could not be processed",
                        "schema": {
                            "$ref": "#/definitions/server.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "color.RGBA": {
            "type": "object",
            "properties": {
                "r": {
                    "type": "integer"
                }
            }
        },
        "server.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "server.Success": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "string"
                }
            }
        },
        "spotify.FullTrack": {
            "type": "object",
            "properties": {
                "album": {
                    "description": "The album on which the track appears. The album object includes a link in href to full information about the album.",
                    "$ref": "#/definitions/spotify.SimpleAlbum"
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/spotify.SimpleArtist"
                    }
                },
                "available_markets": {
                    "description": "A list of the countries in which the track can be played,\nidentified by their ISO 3166-1 alpha-2 codes.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "disc_number": {
                    "description": "The disc number (usually 1 unless the album consists of more than one disc).",
                    "type": "integer"
                },
                "duration_ms": {
                    "description": "The length of the track, in milliseconds.",
                    "type": "integer"
                },
                "explicit": {
                    "description": "Whether or not the track has explicit lyrics.\ntrue =\u003e yes, it does; false =\u003e no, it does not.",
                    "type": "boolean"
                },
                "external_ids": {
                    "description": "Known external IDs for the track.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "external_urls": {
                    "description": "External URLs for this track.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "href": {
                    "description": "A link to the Web API endpoint providing full details for this track.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_playable": {
                    "description": "IsPlayable defines if the track is playable. It's reported when the \"market\" parameter is passed to the tracks\nlisting API.\nSee: https://developer.spotify.com/documentation/general/guides/track-relinking-guide/",
                    "type": "boolean"
                },
                "linked_from": {
                    "description": "LinkedFrom points to the linked track. It's reported when the \"market\" parameter is passed to the tracks listing\nAPI.",
                    "$ref": "#/definitions/spotify.LinkedFromInfo"
                },
                "name": {
                    "type": "string"
                },
                "popularity": {
                    "description": "Popularity of the track.  The value will be between 0 and 100,\nwith 100 being the most popular.  The popularity is calculated from\nboth total plays and most recent plays.",
                    "type": "integer"
                },
                "preview_url": {
                    "description": "A URL to a 30 second preview (MP3) of the track.",
                    "type": "string"
                },
                "track_number": {
                    "description": "The number of the track.  If an album has several\ndiscs, the track number is the number on the specified\nDiscNumber.",
                    "type": "integer"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "spotify.Image": {
            "type": "object",
            "properties": {
                "height": {
                    "description": "The image height, in pixels.",
                    "type": "integer"
                },
                "url": {
                    "description": "The source URL of the image.",
                    "type": "string"
                },
                "width": {
                    "description": "The image width, in pixels.",
                    "type": "integer"
                }
            }
        },
        "spotify.LinkedFromInfo": {
            "type": "object",
            "properties": {
                "external_urls": {
                    "description": "ExternalURLs are the known external APIs for this track or album",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "href": {
                    "description": "Href is a link to the Web API endpoint providing full details",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the linked track",
                    "type": "string"
                },
                "type": {
                    "description": "Type of the link: album of the track",
                    "type": "string"
                },
                "uri": {
                    "description": "URI is the Spotify URI of the track/album",
                    "type": "string"
                }
            }
        },
        "spotify.SimpleAlbum": {
            "type": "object",
            "properties": {
                "album_group": {
                    "description": "The field is present when getting an artist’s\nalbums. Possible values are “album”, “single”,\n“compilation”, “appears_on”. Compare to album_type\nthis field represents relationship between the artist\nand the album.",
                    "type": "string"
                },
                "album_type": {
                    "description": "The type of the album: one of \"album\",\n\"single\", or \"compilation\".",
                    "type": "string"
                },
                "artists": {
                    "description": "A slice of SimpleArtists",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/spotify.SimpleArtist"
                    }
                },
                "available_markets": {
                    "description": "The markets in which the album is available,\nidentified using ISO 3166-1 alpha-2 country\ncodes.  Note that al album is considered\navailable in a market when at least 1 of its\ntracks is available in that market.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "external_urls": {
                    "description": "Known external URLs for this album.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "href": {
                    "description": "A link to the Web API enpoint providing full\ndetails of the album.",
                    "type": "string"
                },
                "id": {
                    "description": "The SpotifyID for the album.",
                    "type": "string"
                },
                "images": {
                    "description": "The cover art for the album in various sizes,\nwidest first.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/spotify.Image"
                    }
                },
                "name": {
                    "description": "The name of the album.",
                    "type": "string"
                },
                "release_date": {
                    "description": "The date the album was first released.  For example, \"1981-12-15\".\nDepending on the ReleaseDatePrecision, it might be shown as\n\"1981\" or \"1981-12\". You can use ReleaseDateTime to convert this\nto a time.Time value.",
                    "type": "string"
                },
                "release_date_precision": {
                    "description": "The precision with which ReleaseDate value is known: \"year\", \"month\", or \"day\"",
                    "type": "string"
                },
                "uri": {
                    "description": "The SpotifyURI for the album.",
                    "type": "string"
                }
            }
        },
        "spotify.SimpleArtist": {
            "type": "object",
            "properties": {
                "external_urls": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "href": {
                    "description": "A link to the Web API enpoint providing full details of the artist.",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uri": {
                    "description": "The Spotify URI for the artist.",
                    "type": "string"
                }
            }
        },
        "spoty.Image": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "hex": {
                    "type": "string"
                },
                "rgba": {
                    "$ref": "#/definitions/color.RGBA"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        }
    }
}